requirements.txt
rasa==3.6.20
python-dotenv==1.0.0
supabase==1.0.3
httpx==0.23.3
spacy==3.7.2 